+++ turbecs/src/assets/prefabs/general_prefabs.rs
@@ -36,11 +36,11 @@ pub fn new_title () -> (Entity, VecDeque<Component>) {
     text_box.font = "large".to_string();
     text_box.color = 0xff0000ff;
 
-    text_box.transform.set_width(TextComponent::get_text_offset(&text_box.text, &text_box.font).get_x() * 4);
+    text_box.transform.set_width(TextComponent::get_text_offset(&text_box.text, &text_box.font).get_x() as i32 * 4);
     text_box.transform.set_height(TextComponent::get_text_offset(&text_box.text, &text_box.font).get_y() as i32 * -2);
     // text_box.transform.set_scale(1.2);
 
-    ent.transform.nudge_y(text_box.transform.get_height() / 2);
+    ent.transform.nudge_y((text_box.transform.get_height() / 2) as f32);
 
     text_box.transform.position.set_horizontal_pref(bound_data::Horizontal::Center);
 
@@ -55,7 +55,7 @@ pub fn new_to_misc() -> (Entity, VecDeque<Component>) {
     let mut ent = Entity::new_base("Misc".to_string());
     let mut ent_queue = VecDeque::new();
 
-    ent.transform.position.nudge_y( -40);
+    ent.transform.position.nudge_y( -40.0);
 
     ent.transform.position.set_vertical_pref(bound_data::Vertical::Bottom);
 
@@ -95,8 +95,8 @@ pub fn new_to_title() -> (Entity, VecDeque<Component>) {
     let mut ent = Entity::new_base("Back".to_string());
     let mut ent_queue = VecDeque::new();
 
-    ent.transform.nudge_y( (screen().h() / 2 * 17 / 20) as i32 );
-    ent.transform.nudge_x((screen().h() / 2 * 3 / 4) as i32 * -1);
+    ent.transform.nudge_y( (screen().h() / 2 * 17 / 20) as f32 );
+    ent.transform.nudge_x((screen().h() / 2 * 3 / 4) as f32 * -1.0);
 
     ent.transform.position.set_vertical_pref(bound_data::Vertical::Bottom);
